requirements.txt
openai
flask
python-dotenv
flask_session
mysql-connector-python